+++ cli/upload.ts
@@ -30,7 +30,7 @@ import {validateMetadata} from './validate';
 
 interface UploadArgs {
   manifestFile: string;
-  codaApiEndpoint: string;
+  apiEndpoint: string;
   notes?: string;
   intermediateOutputDirectory: string;
   timerStrategy: TimerShimStrategy;
@@ -51,7 +51,7 @@ function cleanup(intermediateOutputDirectory: string, logger: Logger) {
 export async function handleUpload({
   intermediateOutputDirectory,
   manifestFile,
-  codaApiEndpoint,
+  apiEndpoint,
   notes,
   timerStrategy,
   apiToken,
@@ -64,9 +64,9 @@ export async function handleUpload({
   }
 
   const manifestDir = path.dirname(manifestFile);
-  const formattedEndpoint = formatEndpoint(codaApiEndpoint);
-  apiToken = assertApiToken(codaApiEndpoint, apiToken);
-  const packId = assertPackId(manifestDir, codaApiEndpoint);
+  const formattedEndpoint = formatEndpoint(apiEndpoint);
+  apiToken = assertApiToken(apiEndpoint, apiToken);
+  const packId = assertPackId(manifestDir, apiEndpoint);
 
   logger.info('Building Pack bundle...');