+++ testing/compile.ts
@@ -5,7 +5,6 @@ import {ensureUnreachable} from '../helpers/ensure';
 import * as esbuild from 'esbuild';
 import exorcist from 'exorcist';
 import fs from 'fs';
-import {getPackOptions} from '../cli/config_storage';
 import {importManifest} from '../cli/helpers';
 import os from 'os';
 import path from 'path';
@@ -142,12 +141,8 @@ function getTimerShims(timerStrategy: TimerShimStrategy): string[] {
   }
 }
 
-function getInjections({timerStrategy = TimerShimStrategy.None, manifestPath}: CompilePackBundleOptions): string[] {
-  const options = getPackOptions(path.dirname(manifestPath));
-  const timerStrategyToUse = options?.timerStrategy || timerStrategy;
-  const shims = [...getTimerShims(timerStrategyToUse), `${__dirname}/injections/crypto_shim.js`];
-
-  return shims;
+function getInjections({timerStrategy = TimerShimStrategy.None}: CompilePackBundleOptions): string[] {
+  return [...getTimerShims(timerStrategy), `${__dirname}/injections/crypto_shim.js`];
 }
 
 async function buildWithES({